Voice-AI-Prompt-Optimizer/requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Voice to prompt
pyaudio>=0.2.13

# Visualization (optional but recommended)
matplotlib>=3.7.0
